Fat 16557 (#4219)
* derived test Spitfire team

* changed test type

* changed test type
