+++ cypress/e2e/inventory/advanced-search/advanced-search-in-inventory.cy.js
@@ -33,22 +33,22 @@ describe('Inventory', () => {
 
     const marcFiles = [
       {
-        marc: 'marcBibFileC400610.mrc',
-        fileName: `testMarcFileC400610.${getRandomPostfix()}.mrc`,
+        marc: 'marcBibFileC466156.mrc',
+        fileName: `C466156 testMarcFile.${getRandomPostfix()}.mrc`,
         jobProfileToRun: DEFAULT_JOB_PROFILE_NAMES.CREATE_INSTANCE_AND_SRS,
         numberOfRecords: 2,
         propertyName: 'instance',
       },
       {
         marc: 'marcBibFileC400616.mrc',
-        fileName: `testMarcFileC400616.${getRandomPostfix()}.mrc`,
+        fileName: `C400616 testMarcFile.${getRandomPostfix()}.mrc`,
         jobProfileToRun: DEFAULT_JOB_PROFILE_NAMES.CREATE_INSTANCE_AND_SRS,
         numberOfRecords: 4,
         propertyName: 'instance',
       },
       {
         marc: 'marcBibFileC414977.mrc',
-        fileName: `testMarcFileC414977.${getRandomPostfix()}.mrc`,
+        fileName: `C414977 testMarcFile.${getRandomPostfix()}.mrc`,
         jobProfileToRun: DEFAULT_JOB_PROFILE_NAMES.CREATE_INSTANCE_AND_SRS,
         numberOfRecords: 9,
         propertyName: 'instance',
