+++ cypress/e2e/marc/marc-authority/update-marc-auth-via-marc-auth-app.cy.js
@@ -90,6 +90,7 @@ describe('MARC', () => {
 
           QuickMarcEditor.pressSaveAndClose();
           cy.wait(1500);
+          cy.pause();
           QuickMarcEditor.pressSaveAndClose();
           QuickMarcEditor.checkCallout(testData.calloutMessage);
           MarcAuthority.contains(
