+++ cypress/e2e/marc/marc-bibliographic/derive-marc-bib/manual-linking/derive-new-marc-bib-unlink-marc-bib-fields-using-remove-linking-button.cy.js
@@ -135,7 +135,7 @@ describe('MARC', () => {
 
       it(
         'C366115 Derive a new MARC bib record: Unlink "MARC Bibliographic" fields from "MARC Authority" records using "Remove linking" button in "Remove authority linking" modal (spitfire) (TaaS)',
-        { tags: ['extendedPath', 'spitfire'] },
+        { tags: ['criticalPath', 'spitfire'] },
         () => {
           InventoryInstances.searchByTitle(createdRecordIDs[0]);
           InventoryInstances.selectInstance();
